backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.1
firebase-admin==6.3.0
python-dotenv==1.0.0
requests==2.31.0